Initial import. This is in a very early stage.
Most things are still handled through plwm. The focus mixin was lifted from
plwm and reworked so that MoveFocus behaves much more sanely so long as the
layout stays perfectly tiles (no gaps whatsoever). There was an issue with the
pane splitting code that offset windows by a pixel or three, breaking focus. I
think that's fixed, but the way I'm set up right now I can't spawn a term in a
new split so I'm not positive.
Overall pretty happy with plwm, but I can see some extra complexity it deals
with to cater to floating wms. I plan to strip these parts out where it bugs me
and possibly even do it with straight Xlib (preferably XCB, actually) once I
understand how this works.
Still a lot of work to do with panes. I need to be able to visibly set the
"focus" to an empty pane, but plwm deals with focus in windows, so I need to
introduce some idea of a fake focus, at least to the point where we can draw a
border around the active pane rather than the active window.